Split IPC module implementation
